refactor: move files from pipelines to cli
