+++ packages/cli/src/commands/pipelines/setup.ts
@@ -5,18 +5,18 @@ import {CliUx} from '@oclif/core'
 
 import Debug from 'debug'
 
-import {createPipeline, getAccountInfo, getTeam} from '../../api'
-import GitHubAPI from '../../github-api'
-import KolkrabbiAPI from '../../kolkrabbi-api'
-import createApps from '../../setup/create-apps'
-import getCISettings from '../../setup/get-ci-settings'
-import getGitHubToken from '../../setup/get-github-token'
-import getNameAndRepo from '../../setup/get-name-and-repo'
-import getRepo from '../../setup/get-repo'
-import getSettings from '../../setup/get-settings'
-import pollAppSetups from '../../setup/poll-app-setups'
-import setupPipeline from '../../setup/setup-pipeline'
-import {nameAndRepo, STAGING_APP_INDICATOR} from '../../setup/validate'
+import {createPipeline, getAccountInfo, getTeam} from 'src/lib/pipelines/api'
+import GitHubAPI from 'src/lib/pipelines/github-api'
+import KolkrabbiAPI from 'src/lib/pipelines/kolkrabbi-api'
+import createApps from '../../lib/pipelines/setup/create-apps'
+import getCISettings from '../../lib/pipelines/setup/get-ci-settings'
+import getGitHubToken from '../../lib/pipelines/setup/get-github-token'
+import getNameAndRepo from '../../lib/pipelines/setup/get-name-and-repo'
+import getRepo from '../../lib/pipelines/setup/get-repo'
+import getSettings from '../../lib/pipelines/setup/get-settings'
+import pollAppSetups from '../../lib/pipelines/setup/poll-app-setups'
+import setupPipeline from '../../lib/pipelines/setup/setup-pipeline'
+import {nameAndRepo, STAGING_APP_INDICATOR} from '../../lib/pipelines/setup/validate'
 
 // eslint-disable-next-line new-cap
 const debug = Debug('pipelines:setup')
