+++ packages/cli/src/commands/pipelines/update.ts
@@ -3,7 +3,7 @@ import {Command, flags} from '@heroku-cli/command'
 import {StageCompletion} from '@heroku-cli/command/lib/completions'
 import {CliUx} from '@oclif/core'
 
-import {updateCoupling} from '../../api'
+import {updateCoupling} from 'src/lib/pipelines/api'
 
 const cli = CliUx.ux
