+++ packages/cli/src/lib/pipelines/infer.ts
