+++ packages/cli/src/lib/pipelines/setup/get-ci-settings.ts
