+++ packages/cli/src/lib/pipelines/setup/get-name-and-repo.ts
