+++ packages/cli/src/lib/pipelines/setup/get-settings.ts
