+++ packages/cli/src/lib/pipelines/setup/setup-pipeline.ts
