+++ packages/cli/src/lib/pipelines/setup/validate.ts
