+++ packages/cli/test/unit/commands/pipelines/add.unit.test.ts
