+++ packages/cli/test/unit/commands/pipelines/connect.unit.test.ts
