+++ packages/cli/test/unit/commands/pipelines/destroy.unit.test.ts
