+++ packages/cli/test/unit/commands/pipelines/diff.unit.test.ts
