+++ packages/cli/test/unit/commands/pipelines/info.unit.test.ts
