+++ packages/cli/test/unit/commands/pipelines/open.unit.test.ts
