+++ packages/cli/test/unit/commands/pipelines/remove.unit.test.ts
