+++ packages/cli/test/unit/commands/pipelines/rename.unit.test.ts
